amirtharaj/azure-python-sdk/modules/requirements.txt

azure-identity 
azure-mgmt-rdbms 
azure-mgmt-rdbms-postgresql-models  
azure-mgmt-rdbms-postgresql_flexibleservers-models 